pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-grievance-routing"
version = "0.1.0"
description = "Grievance Routing environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
server = "grievance_routing.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["grievance_routing", "grievance_routing.server"]
package-dir = { "grievance_routing" = ".", "grievance_routing.server" = "server" }
